Dictionaries to Assign Colors to Grid
uses dictionaries, one with arrays of 1s and 0s representing random data and another dictionary that stores the arrays of the colors corresponding to the first dictionary's data, then displays in 10x10 grid representing a single panel. Then saves the image as png file. Still working to get rid of the white space above the grid.